requirements.txt
gradio==4.43.0
ultralytics==8.3.28
requests==2.32.3
pillow==10.2.0
fastapi==0.112.4
opencv-python==4.9.0.80
numpy==1.26.4
scikit-learn==1.5.2
uvicorn==0.32.1